Port build_faces_from_cells() (first half of build_interior())
Note that the cell type ID is augmented by 1 (the valid type ID started from 0 in SOLVCON, but now it starts from 1 in modmesh).  Use of 1-based index leaves 0 to be the invalid sentinel for the type ID in modmesh.

Use the trivial triangle unit tests for build_faces_from_cells().