Embed Python to viewer
* Add the Qt3DWindow to a QWidget in QMainWindow

After adding the widget, a random SIGBUS (bus error: 10) appeared.  It seems the deleteLater() invoked in ~REntityBase() may trigger double free when REntityBase is allocated on the stack.  To resolve, I remove ~REntityBase().

In the future I should change all R* classes to sub-class from Qt.  Although I do not know whether or not it is correct, many examples does it for Qt so I should learn how to do it.