Correct the wrong deletion of the controller of scene
The camera controller of scene is also a child of it.  For some reason,
QObject::findChildren<RStaticMesh<ND> *>() does not only return RStaticMesh<ND>
*, but also the controller.  Consequently the controller gets deleted when
"show" is called to replace the RStaticMesh.

The fix uses a loop through QNode::childNodes() and RTTI in the loop to
correctly delete only RStaticMesh<ND>.