Remove duplicated edges when drawing the wireframe of a mesh
This is to address issue #68.

A new array `ednds` is added to StaticMesh.  Each row contains the two nodes forming the edge, from the tail to head.

A new (private) argument _build_edge is added to StaticMesh.build_interior() and its default value is False.  A standalone function build_edge() is added to StaticMesh too.

RStaticMesh::update_geometry_impl() is updated to use the new ednds array.

A new free helper function makeQByteArray() is added to copy the data from a SimpleArray to Qt QByteArray.