Refactor: Embed the base Python extension module to the viewer executable
The "base" _modmesh Python extension module was built in a separate shared object and loaded into the viewer binary.  Some object files need to be linked in the two images.  It is unnecessarily complex.  Embedding the "base" Python extension module in the same way that the "_modmesh_view" module is embedded to the viewer executable simplifies the linkage.

The pure Python module (package) `modmesh` is changed accordingly.  When it is used without the Qt viewer, it still tries to load the `_modmesh` extension module relatively.