Fix viewer crash with linear_wave app loaded
The linear_wave app uses PySide6 for matplotlib to create a new window, and it seems to conflict with the shutdown sequence of python interpreter.

Change the interpreter shotdown order to fix it.

In addition, improve the interpreter wrapping API.  Do not initialize the interpreter in the constructor of the class Interpreter.