Create a new Python module `modmesh.apputil` to manage app environment
Remove the primitive C++ code for running the viewer pytext code for an application to the new module `modmesh.apputil`.  A new class `AppEnvironment` is defined to hold the symbols for the application, so that the symbols will not scatter in the interpreter `__main__`.

Each run of the applications will create a new app environment, which is collected in the list `modmesh.apputil.environ`.

Also in this commit: Rename `FIXME` to `TODO`.