Make the viewer binary a multi-purpose binary
Change the start-up code so that the viewer binary can run in the following 4 modes:
1. viewer: open a windows
2. command: take a string as Python code
3. terminal: start text-based iPython prompt
4. pytest: run the Python test cases

In this way, the viewer binary can be used to run the Python test cases.  A new target run_viewer_pytest is added to the make file.  CI (Ubuntu, macos, and Windows) uses the new target to run unit tests.