Add an interactive Python command console
Add a dock widget at the bottom of the main window that a horizontally split text edit areas.  The upper and larger one is for displaying the history of the input Python commands.  The lower and narrow one is to input a Python command for execution.

The command line supports history navigation through up and down arrow keys on the keyboard, and uses enter for sending the command for execution.  It does not yet support multi-line Python code.

Improvements that are needed in the future:
* The output from Python goes to the process stdout and stderr that are not redirected to the Qt widget.  It should be in the future.
* The Python code dock widget on the right-hand side of the main window will be kept for a while before all the apps are migrated to use the new Python command console.