Revert fa21d6e18904df4c6c05f3d93c7f911e90f7b6f4
I was wrong in the review https://github.com/solvcon/modmesh/pull/167#pullrequestreview-1211075520 to think it is ok to set vertex size to an illegal value.  It is not because it may result into memory error.  I think it is implied in the qt3d doc: https://doc.qt.io/qt-6/qt3dcore-qattribute.html#vertexSize-prop.

I will file an issue to track it.  The vertex size may be an important parameter for qt3d and I do not fully understand its meaning.