Add `__repr__` and `__str__` for `Point3d`, `Segment3d`, `Bezier3d`
Implement the string representing helpers `__repr__` and `__str__` for the Python classes `Point3d`, `Segment3d`, and `Bezier3d` (both the `Fp32` and `Fp64` flavors).  It helps analyze the graphical objects in the "World" object in the modmesh pilot command-line window.

Also add the objects `world` and `wid` of the SVG app to the appenv (application environment) so that users can access them from the command line.